+++ Common/Integrations/GenericModConfigMenu/GenericModConfigMenuIntegration.cs
@@ -265,6 +265,15 @@ internal static class GenericModConfigMenuIntegration
 
         GenericModConfigMenuIntegration<TConfig> api = new(mod.Helper.ModRegistry, mod.Monitor, mod.ModManifest, get, Reset, SaveAndApply);
         if (api.IsLoaded)
-            configMenu.Register(api, mod.Monitor);
+        {
+            try
+            {
+                configMenu.Register(api, mod.Monitor);
+            }
+            catch (Exception ex)
+            {
+                mod.Monitor.LogOnce($"Failed registering config menu with Generic Mod Config Menu.\n\nTechnical info:\n{ex}", LogLevel.Error);
+            }
+        }
     }
 }