+++ LookupAnything/Framework/Lookups/Characters/FarmerSubject.cs
@@ -75,8 +75,8 @@ internal class FarmerSubject : BaseSubject
         var spaceCore = this.GameHelper.SpaceCore;
         if (spaceCore.IsLoaded)
         {
-            foreach (string skill in spaceCore.ModApi.GetCustomSkills())
-                yield return new SkillBarField(spaceCore.ModApi.GetDisplayNameOfCustomSkill(skill), spaceCore.ModApi.GetExperienceForCustomSkill(target, skill), maxSkillPoints, skillPointsPerLevel);
+            foreach (string skill in spaceCore.GetCustomSkills())
+                yield return new SkillBarField(spaceCore.GetDisplayNameOfCustomSkill(skill), spaceCore.GetExperienceForCustomSkill(target, skill), maxSkillPoints, skillPointsPerLevel);
         }
 
         // luck