+++ Automate/docs/release-notes.md
@@ -1,7 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
-## Upcoming release
+## 2.4.4
+Released 26 October 2025 for SMAPI 4.1.10 or later.
+
 * Fixed some machine stats not updated correctly.
 * Fixed features breaking if an integration with another mod fails.